+++ src/vimeotools/vimeo_video.py
@@ -15,7 +15,8 @@ from typing import TYPE_CHECKING, Dict, Optional, List, Union, Any, Literal
 from vimeo_base import (
     VimeoChild,
     VimeoItem,
-    get_lines
+    get_lines,
+    transform_returning
 )
 from vimeo_constants import (
     GETTER_STR,
@@ -29,6 +30,7 @@ if TYPE_CHECKING:
     from vimeo_connection import VimeoConnection
     from vimeo_showcase import VimeoShowcase
     from vimeo_folder import VimeoFolder
+    from vimeo_data import VimeoData
 
 class VimeoVideo(VimeoItem, VimeoChild):
     """
@@ -39,13 +41,15 @@ class VimeoVideo(VimeoItem, VimeoChild):
 
     for key, val in PROPERTIES_VIDEO.items():
         exec(GETTER_STR.format(prop=key))
-        if val['type'] == 'str' and val.get('setable'):
+
+        if val['type'] == str and val.get('setable'):
             set_key = val.get('set_key')
             if set_key is None:
                 continue  # needs a special setter
-            key = set_key
-            exec(SETTER_STR.format(prop=set_key))
-    
+
+            setter_cmd = SETTER_STR.format(prop=key, set_key=set_key)
+            exec(setter_cmd)
+
     @property
     def tags(self) -> List[str]:
         return [tag['name'] for tag in self.get_property('tags')]
@@ -57,12 +61,16 @@ class VimeoVideo(VimeoItem, VimeoChild):
     def __init__(
         self,
         connection: 'VimeoConnection',
-        code_or_uri: str
+        code_or_uri: str,
+        data: Optional[Dict[str, Any]] = None,
+        data_object: Optional['VimeoData'] = None
     ):
         # init the parent class
         super(VimeoItem, self).__init__(
             connection=connection,
-            code_or_uri=code_or_uri
+            code_or_uri=code_or_uri,
+            data=data,
+            data_object=data_object
         )
 
         super(VimeoChild, self).__init__() # type: ignore
@@ -139,6 +147,154 @@ class VimeoVideo(VimeoItem, VimeoChild):
         else:
             raise Exception(f'Error: {response.status_code}')
 
+    def add_tag(
+        self,
+        tag: str,
+        refresh: bool = False
+    ):
+        self.add_tags(
+            tags=tag,
+            refresh=refresh
+        )
+
+    def add_tags(
+        self,
+        tags: Union[str, List[str]],
+        refresh: bool = False
+    ):
+        """
+        Add tags to the video.
+        :param tags: str or list of str
+        :param refresh: bool
+        :return: None
+        """
+        if isinstance(tags, str):
+            tags = [tags]
+        
+        tags = self.get_tags(refresh=refresh) + tags # type: ignore
+
+        self.set_tags(
+            tags=tags
+        )
+
+    def get_tags(
+        self,
+        returning: Literal[
+            'dict',
+            'list',
+            'json',
+            'names',
+            'name',
+            'uris',
+            'uri',
+            'tags',
+            'tag',
+            'canonical'
+        ] = 'dict',
+        refresh: bool = False
+    ) -> Union[List[str], str]:
+        tags = self.get_property('tags', refresh=refresh)
+
+        # not sure why data has not always the same structure
+        if isinstance(tags, dict):
+            tags = tags['data']
+
+        if returning == 'dict':
+            return tags
+        elif returning == 'list':
+            return tags['data']
+        elif returning == 'json':
+            return json.dumps(tags)
+        else:
+            if returning.endswith('s'):
+                returning = returning[:-1] # type: ignore
+            return [tag[returning] for tag in tags]
+
+    def remove_tag(
+        self,
+        tag: str
+    ):
+        """
+        Remove a tag from the video.
+        :param tag: str
+        :return: None
+
+        This method uses the delete method of the Vimeo API.
+        """
+        response = self.client.delete(
+            f'{self.uri}/tags/{tag}'
+        )
+        if not response.status_code == 204:
+            raise Exception(f'Error: {response.status_code}')
+        
+        self._data['tags'] = [tag for tag in self._data['tags'] if tag != tag]
+
+    def remove_tags(
+        self,
+        tags: Union[str, List[str]],
+        refresh: bool = False
+    ):
+        """
+        Remove tags from the video.
+        :param tags: str or list of str
+        :param refresh: bool
+        :return: None
+
+        This method uses the put method of the Vimeo API.
+        """
+        if isinstance(tags, str):
+            tags = [tags]
+        
+        tags_list = [{'name': tag} for tag in self.get_tags(refresh=refresh) if tag not in tags]
+        
+        response = self.client.put(
+            f'{self.uri}/tags',
+            data=tags_list
+        )
+        assert response.status_code == 200, f'Error: {response.status_code}'
+        self._data['tags'] = tags  # type: ignore
+
+    def set_description(self, value: str):
+        self.set_property(
+            key='description',
+            value=value
+        )
+
+    def set_tags(
+        self,
+        tags: Union[str, List[str], List[Dict[str, str]]]
+    ):
+        set_tags = []
+        for tag in tags:
+            if isinstance(tag, str):
+                set_tags.append({'name': tag})
+            if isinstance(tag, dict):
+                set_tags.append({'name': tag['name']})
+
+        uri = f'{self.uri}/tags'
+        
+        tags_list = [tag for tag in set_tags]
+        response = self.client.put(
+            uri,
+            data=tags_list
+        )
+        if response.status_code == 400:
+            raise Exception((
+                '400: Bad Request,'
+                "either the request body wasn't supplied,"
+                'or a parameter is invalid, '
+                "The request body doesn't contain a JSON-encoded list of tags. "
+                f'You supplied: {tags_list}'
+            ))
+        elif response.status_code == 403:
+            raise Exception((
+                '403: Forbidden, you are not allowed to add tags to this video.'
+                'You may not have the permission or the total number of tags exceeds the limit (20). '
+                f'You supplied: {tags_list}'
+            ))
+        assert response.status_code == 200, f'Error: {response.status_code}'
+        self._data['tags'] = response.json()  # type: ignore
+        
     def get_chapter(
         self,
         chapter_id: str,
@@ -223,7 +379,18 @@ class VimeoVideo(VimeoItem, VimeoChild):
 
     def get_showcases(
         self,
-        returning: Literal['obj', 'list', 'dict', 'names', 'codes', 'uris'] = 'obj',
+        returning: Literal[
+            'object',
+            'objects',
+            'list',
+            'dict',
+            'name',
+            'names',
+            'code',
+            'codes',
+            'uri',
+            'uris'
+        ] = 'object',
         refresh: bool = False
     ) -> Union[List[str], Dict[str, Any], List['VimeoShowcase']]:
         """
@@ -232,6 +399,9 @@ class VimeoVideo(VimeoItem, VimeoChild):
         :param refresh: bool
         :return: Union[List[str], Dict[str, Any]]
         """
+        returning = transform_returning(returning)
+        refresh = refresh or self._live
+
         if self._showcases is None or refresh:
             response = self.client.get(f'/videos/{self.code}/albums')
             assert response.status_code == 200
@@ -240,7 +410,7 @@ class VimeoVideo(VimeoItem, VimeoChild):
         else:
             videos_d = self._showcases
 
-        if returning == 'obj':
+        if returning == 'object':
             return [
                 VimeoShowcase(
                     showcase['uri'].split('/')[-1],
@@ -253,14 +423,11 @@ class VimeoVideo(VimeoItem, VimeoChild):
             return videos_d
         elif returning == 'list':
             return videos_d['data']
-        elif returning == 'names':
-            return [video['name'] for video in videos_d['data']]
-        elif returning == 'codes':
-            return [video['uri'].split('/')[-1] for video in videos_d['data']]
-        elif returning == 'uris':
-            return [video['uri'] for video in videos_d['data']]
         else:
-            raise ValueError(f'Invalid returning value: {returning}')
+            try:
+                return [video[returning] for video in videos_d['data']]
+            except KeyError:
+                raise ValueError(f'Invalid returning value: {returning}')
     
     @property
     def status(